backend/requirements.txt
# arxiv
pandas
agno
# llama-index
openai
# sentence-transformers[openvino]
fastapi[standard]
python-dotenv
chromadb